modified existing tests, and added independent file for buttons
